# docs/swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "Spacebox API",
    "description": "for spacebox",
    "version": "1.0"
  },
  "produces": ["application/json", "text/plain"],
  "host": "spacebox-tech.herokuapp.com",
  "basePath": "/",
  "definitions": {
...
    "uuid": {
      "type": "string",
      "pattern" : "^[0-9a-f]{8}-[0-9a-f]{4}-[0-9a-f]{4}-[0-9a-f]{4}-[0-9a-f]{12}$"
    },
    "responseError": {
      "type": "object",
      "properties": {
        "errorDetails": {
          "description": "this really should be rationalized but atm, it could be a string OR an object. Swagger doesn't support this as a schema so it's left undefined"
        }
      }
    },
...
    "shootingParameters": {
      "type": "object",
      "properties": {
        "vessel": {
          "description": "the uuid of the ship to shoot from",
          "$ref": "#/definitions/uuid"
        },
        "target": {
          "description": "the uuid of the target to shoot",
          "$ref": "#/definitions/uuid"
        }
      },
      "required": [ "vessel", "target" ]
    },
...
    "deployParameters": {
      "type": "object",
      "properties": {
        "container_id": {
          "description": "the uuid of the container to deploy from",
          "$ref": "#/definitions/uuid"
        },
        "slice": {
          "description": "the slice to deploy from",
          "type": "string",
          "default": "default"
        },
        "blueprint": {
          "description": "the uuid of the blueprint to deploy an item of",
          "$ref": "#/definitions/uuid"
        },
        "vessel_uuid": {
          "description": "the uuid of vessel to deploy if a specific one already exists",
          "$ref": "#/definitions/uuid"
        }
      },
      "required": [ "container_id", "blueprint" ]
    },
...
  },
  "securityDefinitions": {
      "swagger-ui-key": {
        "description": "This just defines what the Swagger-UI does so it has to be supported. The preferred method is to pass in the JWT token as a bearer token in the Authorization header.",
        "name": "api_key",
        "type": "apiKey",
        "in": "query"
      }
  },
  "security": [
    {
      "swagger-ui-key": []
    }
  ],
  "responses": {
    "Error": {
      "description": "Any kind of error, not just an internal error",
      "schema": {
        "$ref": "#/definitions/responseError"
      }
    },
...
  },
  "paths": {
...
    "/commands/shoot": {
      "post": {
        "tags": [ "spaceObjects" ],
        "description": "Tell a vessel to shoot a target until it's destroyed",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "description": "the parameters as json",
            "schema": {
              "$ref": "#/definitions/shootingParameters"
            }
          }
        ],
        "responses": {
          "default": { "$ref": "#/responses/Error" },
          "200": {
            "description": "success",
            "schema": {
              "type": "object",
              "properties": {
                "result": {
                  "type": "boolean",
                  "enum": [ true ]
                }
              }
            }
          }
        }
      }
    },
    "/commands/deploy": {
      "post": {
        "tags": [ "spaceObjects" ],
        "description": "deploy a vessel from inventory to space",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "description": "the parameters as json",
            "schema": {
              "$ref": "#/definitions/deployParameters"
            }
          }
        ],
        "responses": {
          "default": { "$ref": "#/responses/Error" },
          "200": {
            "description": "success",
            "schema": {
              "type": "object",
              "properties": {
                "result": {
                  "description": "the uuid of the new vessel",
                  "$ref": "#/definitions/uuid"
                }
              }
            }
          }
        }
      }
    },
...
    "/commands/spawnStarter": {
      "post": {
        "tags": [ "spaceObjects" ],
        "description": "spawns a starter ship, only works if you have no game assets",
        "parameters": [],
        "responses": {
          "default": { "$ref": "#/responses/Error" },
          "200": {
            "description": "success",
            "schema": {
              "type": "object",
              "properties": {
                "result": {
                  "description": "the uuid of the starter ship",
                  "$ref": "#/definitions/uuid"
                }
              }
            }
          }
        }
      }
    },
...
  }
}